+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Automatic addition of device ID as metadata to shots after saving them
 - New config option `DEVICE_TYPE` to define the device to run on
 - Add OpenAPI Specification (OAS) using Swagger under `/api`
+- Introduce settings `picture quality` and `movie quality`
 
 ### Changed